+++ BasketProject/Products/MilkProduct.cs
@@ -2,6 +2,6 @@ namespace BasketProject.Products
 {
     public class MilkProduct : ProductBase
     {
-        public override double OriginalPrice => 1.15;
+        public override decimal OriginalPrice => 1.15M;
     }
 }
