+++ BasketProject/Products/ProductBase.cs
@@ -2,10 +2,10 @@ namespace BasketProject.Products
 {
     public abstract class ProductBase
     {
-        abstract public double OriginalPrice {get;}
+        abstract public decimal OriginalPrice {get;}
 
-        public double? Dicount {get;set;}
+        public decimal? Dicount {get;set;}
 
-        public double Price => Dicount.HasValue ? OriginalPrice - (OriginalPrice * (Dicount.Value / 100)) : OriginalPrice;
+        public decimal Price => Dicount.HasValue ? OriginalPrice - (OriginalPrice * (Dicount.Value / 100)) : OriginalPrice;
     }
 }
