+++ BasketUnitTests/TestProducts/TestProduct1.cs
@@ -4,6 +4,6 @@ namespace BasketUnitTests.TestProducts
 {
     public class TestProduct1 : ProductBase
     {
-        public override double OriginalPrice => 3.50;
+        public override decimal OriginalPrice => 3.50M;
     }
 }
